backend/requirements.txt
asgiref==3.6.0
astroid==2.14.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
dill==0.3.6
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-filter==22.1
django-js-asset==2.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
pipreqs==0.4.11
platformdirs==3.0.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.16.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.14
whitenoise==6.3.0
wrapt==1.14.1
yarg==0.1.9
